requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic-settings==2.3.4

# ── LangChain ecosystem — semua harus satu family versi ──────────────────────
# Upgrade ke versi terbaru agar langchain-groq kompatibel
langchain>=0.3.0
langchain-core>=1.0.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-groq>=1.1.0

# ── Storage & Document ────────────────────────────────────────────────────────
faiss-cpu==1.8.0
pypdf==4.2.0